Fix #1176: [Bug, plots] DRY violations in ARIMA feed plots
