+++ pdr_backend/statutil/test/test_autocorrelation.py
@@ -1,12 +1,13 @@
 from enforce_typing import enforce_types
 import numpy as np
 import pandas as pd
+import pytest
 
-from pdr_backend.statutil.autocorrelation import (
+from pdr_backend.statutil.autocorrelation_plotdata import (
     AutocorrelationPlotdata,
     AutocorrelationPlotdataFactory,
-    plot_autocorrelation,
 )
+from pdr_backend.statutil.autocorrelation_plotter import plot_acf, plot_pacf
 from pdr_backend.statutil.boxcox import safe_boxcox
 
 DATA_FILE = (
@@ -24,14 +25,20 @@ def test_autocorrelation_SHOW_PLOT():
 
 
 @enforce_types
-def test_autocorrelation():
-    # play with me
-    nlags = 5
-    do_boxcox = True
-    differencing_order = 1
+@pytest.mark.parametrize(
+    "do_boxcox, differencing_order",
+    [
+        (False, 0),
+        (True, 0),
+        (True, 1),
+        (True, 2),
+    ],
+)
+def test_autocorrelation(do_boxcox, differencing_order):
+    nlags = 5  # play with me
 
-    #
-    y = np.array(_get_data())  # [:1000]
+    y = _get_data()
+    y = np.array(y)
     if do_boxcox:
         y = safe_boxcox(y)
     for _ in range(differencing_order):
@@ -40,8 +47,11 @@ def test_autocorrelation():
     d = AutocorrelationPlotdataFactory.build(y, nlags)
     assert isinstance(d, AutocorrelationPlotdata)
 
-    fig = plot_autocorrelation(d)
+    fig = plot_acf(d)
+    if SHOW_PLOT:
+        fig.show()
 
+    fig = plot_pacf(d)
     if SHOW_PLOT:
         fig.show()
